+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support for Oracle distribution of Hadoop with Big Data service
 - Support for APM Synthetics service added
 - Support for Regional VLANs added to `core`
+- Support added for VCN DNS Namespace Collision Remediation
 
 ## 4.34.0 (July 07, 2021)